style: 초대 코드 버튼 위치 변경
