+++ frontend/src/pages/ChatRoom.jsx
@@ -7,6 +7,7 @@ import 'highlight.js/styles/github.css';
 import Sidebar from '../components/SideBar';
 import Header from '../components/header';
 import SearchSidebar from '../components/SearchSideBar';
+import { FaCopy } from 'react-icons/fa';
 
 const ChatRoom = () => {
   const [messages, setMessages]=useState([]);
@@ -17,6 +18,9 @@ const ChatRoom = () => {
   const stompClientRef = useRef(null);
   const messagesEndRef = useRef(null);
   const isComposingRef = useRef(false);
+
+  // 초대 코드 관련 상태 추가
+  const [showNotification, setShowNotification] = useState(false);
 
   const [showSearchSidebar, setShowSearchSidebar] = useState(false);
   const [searchKeyword, setSearchKeyword] = useState('');
@@ -88,6 +92,27 @@ const ChatRoom = () => {
   useEffect(() => {
     scrollToBottom();
   }, [messages]);
+
+  // 초대 코드 복사 기능 추가
+  const copyInviteUrl = async () => {
+    if (!roomId) return; // roomId가 없는 경우 실행하지 않음
+    
+    try {
+      const res = await fetch(`http://localhost:8080/chat-rooms/invite/${roomId}`, {
+        method: 'GET',
+        headers: { 'Content-Type': 'application/json' },
+        credentials: 'include'
+      });
+      if (!res.ok) throw new Error('초대 URL을 가져오지 못했습니다.');
+      const { inviteUrl } = await res.json();
+      await navigator.clipboard.writeText(inviteUrl);
+      setShowNotification(true); // 알림 표시
+      setTimeout(() => setShowNotification(false), 2000); // 2초 뒤 닫기
+    } catch (err) {
+      console.error(err);
+      alert('초대 URL 복사 중 오류가 발생했습니다.');
+    }
+  };
 
   const sendMessage = (text = content) => {
   const trimmed = text.trim();
@@ -331,12 +356,40 @@ const handleSearch = async (keyword, page = 0) => {
             backgroundColor: '#fff'
           }}>
             <div style={{ 
-              fontWeight: '600', 
-              fontSize: '18px',
-              color: '#2d3748'
+              display: 'flex',
+              alignItems: 'center',
+              gap: '12px'
             }}>
-              채팅방 #{roomId}
+              <span style={{ 
+                fontWeight: '600', 
+                fontSize: '18px',
+                color: '#2d3748'
+              }}>
+                채팅방 #{roomId}
+              </span>
+              
+              {/* 초대 코드 복사 버튼 */}
+              <button
+                onClick={copyInviteUrl}
+                style={{
+                  backgroundColor: '#2588F1',
+                  color: 'white',
+                  border: 'none',
+                  borderRadius: '4px',
+                  padding: '6px 12px',
+                  fontSize: '13px',
+                  fontWeight: '500',
+                  cursor: 'pointer',
+                  display: 'flex',
+                  alignItems: 'center',
+                  gap: '6px'
+                }}
+              >
+                <FaCopy size={14} />
+                초대 코드 복사
+              </button>
             </div>
+            
             <div>
               <input
                 type="text"
@@ -516,6 +569,22 @@ const handleSearch = async (keyword, page = 0) => {
           />
         )}
       </div>
+
+      {showNotification && (
+        <div style={{
+          position: 'fixed',
+          top: '15px',
+          right: '15px',
+          backgroundColor: '#333',
+          color: '#fff',
+          padding: '10px 16px',
+          borderRadius: '6px',
+          boxShadow: '0 4px 8px rgba(0, 0, 0, 0.2)',
+          zIndex: 1000
+        }}>
+          초대 코드가 복사되었습니다
+        </div>
+      )}
     </div>
   );
 };
